[typescript] wip
